Works on issue 188 by adding debugging code to the MIDI import.  A vulnerability in eof_get_ts() was identified when I ran EOF through GDB via the command line (instead of CodeBlocks) and I corrected that.  Now I cannot reproduce the crash when I run in debug mode in CodeBlocks or launch it from GDB, but if I run it outside of any debugger, the program crashes somewhere near the end of the beat creation loop.

+++ trunk/src/song.h
@@ -262,13 +262,13 @@ typedef struct
 	EOF_VOCAL_TRACK * vocal_track;
 
 	EOF_BEAT_MARKER * beat[EOF_MAX_BEATS];
-	int beats;
+	unsigned long beats;
 
 	EOF_TEXT_EVENT * text_event[EOF_MAX_TEXT_EVENTS];
-	int text_events;
+	unsigned long text_events;
 
 	/* miscellaneous */
-	int bookmark_pos[EOF_MAX_BOOKMARK_ENTRIES];
+	unsigned long bookmark_pos[EOF_MAX_BOOKMARK_ENTRIES];
 	EOF_CATALOG * catalog;
 
 } EOF_SONG;
